infrastructure/files/cloud_functions/vertex_pipeline/requirements.txt
google-api-python-client
google-cloud-scheduler
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
google-cloud-aiplatform
google-cloud-aiplatform[pipelines]
